fix(frontend): Path for Auto Form
