+++ apps/docs/content/docs/dev/lint.mdx
@@ -1,0 +1,4 @@
+---
+title: Lint
+description: Lint your code with ESLint and Prettier.
+---
