+++ apps/docs/content/docs/dev/meta.json
@@ -5,8 +5,9 @@
     "---Introduction ---",
     "index",
     "contribution",
-    "--- Plugins ---",
+    "--- Plugins in AdminCP ---",
     "plugins",
+    "lint",
     "plugins-nav-admin",
     "plugins-permissions-admin",
     "--- GraphQL ---",
