+++ apps/docs/content/docs/dev/pages.mdx
@@ -3,19 +3,47 @@ title: Pages
 description: Create pages in VitNode.
 ---
 
+Some folders in `apps/frontend/app` are in `.gitignore` and are not included in the repository. All files from these folders are copied when you are using `npm run config:init` to initialize your workspace.
+
+<Callout title="Default pages from VitNode" type="error">
+  Do not put any files inside `(vitnode)` folders! They are automatically
+  generated by VitNode and can be overwritten.
+</Callout>
+
+Below is guide how to create own pages in VitNode using own `plugin`. If your `workspace` is `private` and you don't plan to share your `workspace` with others, you can create pages without `{your_plugin}` folder.
+
 ## Default Page
 
-VitNode handle automaticly default page for you. You need to create a file `themes/1/{your_plugin}/default-page.tsx` in your plugin template.
+VitNode automaticly create default page for you in your plugin.
 
-Use `export default` for your template.
+import { File, Folder, Files } from 'fumadocs-ui/components/files';
 
-```tsx
-export default async function DefaultPage() {
-  const data = await getData();
-  const t = await getTranslations('core');
+<Files>
+  <Folder name="app" defaultOpen>
+    <Folder name="frontend" defaultOpen>
+      <Folder name="plugins" defaultOpen>
+        <Folder name="{your_plugin}" defaultOpen>
+          <Folder name="templates" defaultOpen>
+            <File name="default-page.tsx" />
+          </Folder>
+        </Folder>
+      </Folder>
+    </Folder>
+  </Folder>
+</Files>
 
-  return <div>{t('info', { data })}</div>;
-}
-```
+<Callout title="export default">
+  The default page is exported as `default`.
+</Callout>
 
 Test it by set your plugin as default in `AdminCP` and go to your website.
+
+You can also delete this file if you don't need it, then VitNode after export your plugin will block your plugin to be default.
+
+## Create Pages
+
+Vitnode using [NextJS](https://nextjs.org/docs/app/building-your-application/routing) in app directory. All features about `routing` from `NextJS` are available in VitNode for example [Parallel Routes](https://nextjs.org/docs/app/building-your-application/routing/parallel-routes), [Dynamic Routes](https://nextjs.org/docs/app/building-your-application/routing/dynamic-routes), [Error Handling](https://nextjs.org/docs/app/building-your-application/routing/error-handling), [Metadata](https://nextjs.org/docs/app/building-your-application/optimizing/metadata) etc.
+
+You need to remember only where to put your pages if you are creating plugin for VitNode.
+
+### Frontend
