+++ apps/docs/content/docs/ui/forms/color-picker.mdx
@@ -11,8 +11,8 @@ import { Tab, Tabs } from 'fumadocs-ui/components/tabs';
   <Tab value="Auto Form">
 ```ts
 import * as z from 'zod';
-import { AutoForm } from 'vitnode-frontend/components/ui/auto-form';
-import { AutoFormColor } from 'vitnode-frontend/components/ui/auto-form/fields/color';
+import { AutoForm } from 'vitnode-frontend/form/auto-form';
+import { AutoFormColor } from 'vitnode-frontend/form/fields/color';
 ```
 
 ```ts
