+++ apps/docs/content/docs/ui/forms/combobox.mdx
@@ -16,8 +16,8 @@ import { Tab, Tabs } from 'fumadocs-ui/components/tabs';
   <Tab value="Auto Form">
 ```ts
 import * as z from 'zod';
-import { AutoForm } from 'vitnode-frontend/components/ui/auto-form';
-import { AutoFormCombobox } from 'vitnode-frontend/components/ui/auto-form/fields/combobox';
+import { AutoForm } from 'vitnode-frontend/form/auto-form';
+import { AutoFormCombobox } from 'vitnode-frontend/form/fields/combobox';
 ```
 
 ```ts
