+++ apps/docs/content/docs/ui/forms/file-input.mdx
@@ -11,8 +11,8 @@ import { Tab, Tabs } from 'fumadocs-ui/components/tabs';
   <Tab value="Auto Form">
 ```ts
 import * as z from 'zod';
-import { AutoForm } from 'vitnode-frontend/components/ui/auto-form';
-import { AutoFormFile } from 'vitnode-frontend/components/ui/auto-form/fields/file';
+import { AutoForm } from 'vitnode-frontend/form/auto-form';
+import { AutoFormFile } from 'vitnode-frontend/form/fields/file';
 import { zodFile } from 'vitnode-frontend/helpers/zod';
 ```
 
@@ -62,7 +62,7 @@ import { FileInput } from 'vitnode-frontend/components/ui/file-input';
   <Tab value="Auto Form">
 ```ts
 import * as z from 'zod';
-import { AutoForm } from 'vitnode-frontend/components/ui/auto-form';
+import { AutoForm } from 'vitnode-frontend/form/auto-form';
 import { zodFile } from 'vitnode-frontend/helpers/zod';
 ```
