+++ apps/docs/content/docs/ui/forms/icon-picker.mdx
@@ -11,8 +11,8 @@ import { Tab, Tabs } from 'fumadocs-ui/components/tabs';
   <Tab value="Auto Form">
 ```ts
 import * as z from 'zod';
-import { AutoForm } from 'vitnode-frontend/components/ui/auto-form';
-import { AutoFormIcon } from 'vitnode-frontend/components/ui/auto-form/fields/icon';
+import { AutoForm } from 'vitnode-frontend/form/auto-form';
+import { AutoFormIcon } from 'vitnode-frontend/form/fields/icon';
 ```
 
 ```ts
