+++ apps/docs/content/docs/ui/forms/radio-group.mdx
@@ -16,8 +16,8 @@ import { Tab, Tabs } from 'fumadocs-ui/components/tabs';
   <Tab value="Auto Form">
 ```ts
 import * as z from 'zod';
-import { AutoForm } from 'vitnode-frontend/components/ui/auto-form';
-import { AutoFormRadioGroup } from 'vitnode-frontend/components/ui/auto-form/fields/radio-group';
+import { AutoForm } from 'vitnode-frontend/form/auto-form';
+import { AutoFormRadioGroup } from 'vitnode-frontend/form/fields/radio-group';
 ````
 
 ```ts
