+++ apps/docs/content/docs/ui/forms/tags-input.mdx
@@ -11,8 +11,8 @@ import { Tab, Tabs } from 'fumadocs-ui/components/tabs';
   <Tab value="Auto Form">
 ```ts
 import * as z from 'zod';
-import { AutoForm } from 'vitnode-frontend/components/ui/auto-form';
-import { AutoFormTags } from 'vitnode-frontend/components/ui/auto-form/fields/tags';
+import { AutoForm } from 'vitnode-frontend/form/auto-form';
+import { AutoFormTags } from 'vitnode-frontend/form/fields/tags';
 import { zodTag } from 'vitnode-frontend/helpers/zod';
 ```
