+++ apps/docs/content/docs/ui/forms/text-language-input.mdx
@@ -12,8 +12,8 @@ import { Tab, Tabs } from 'fumadocs-ui/components/tabs';
 ```ts
 import * as z from 'zod';
 import { zodLanguageInput } from 'vitnode-frontend/helpers/zod';
-import { AutoForm } from 'vitnode-frontend/components/ui/auto-form';
-import { AutoFormTextLanguageInput } from 'vitnode-frontend/components/ui/auto-form/fields/text-language-input';
+import { AutoForm } from 'vitnode-frontend/form/auto-form';
+import { AutoFormTextLanguageInput } from 'vitnode-frontend/form/fields/text-language-input';
 ```
 
 ```ts
