+++ apps/docs/content/docs/ui/forms/textarea.mdx
@@ -16,7 +16,7 @@ import { Tab, Tabs } from 'fumadocs-ui/components/tabs';
   <Tab value="Auto Form">
 ```ts
 import * as z from 'zod';
-import { AutoForm } from 'vitnode-frontend/components/ui/auto-form';
+import { AutoForm } from 'vitnode-frontend/form/auto-form';
 ```
 
 ```ts
@@ -58,7 +58,7 @@ import { Textarea } from 'vitnode-frontend/components/ui/textarea';
   <Tab value="Auto Form">
 ```ts
 import * as z from 'zod';
-import { AutoForm } from 'vitnode-frontend/components/ui/auto-form';
+import { AutoForm } from 'vitnode-frontend/form/auto-form';
 ```
 
 ```ts
