+++ packages/frontend/src/components/form/fields/checkbox.tsx
@@ -1,12 +1,12 @@
 import { AutoFormInputComponentProps } from '../type';
-import { AutoFormLabel } from './common/label';
-import { AutoFormTooltip } from './common/tooltip';
 import { cn } from '@/helpers/classnames';
+
+import { FormControl, FormMessage } from '../../ui/form';
+import { Checkbox } from '../../ui/checkbox';
 import { AutoFormWrapper } from './common/wrapper';
 import { DefaultParent } from './common/children';
-
-import { FormControl, FormMessage } from '../../form';
-import { Checkbox } from '../../checkbox';
+import { AutoFormLabel } from './common/label';
+import { AutoFormTooltip } from './common/tooltip';
 
 export const AutoFormCheckbox = ({
   autoFormProps: { isRequired, fieldConfigItem, field, theme, isDisabled },
