+++ packages/frontend/src/components/form/fields/file.tsx
@@ -1,21 +1,16 @@
+import { FileInput } from '@/components/ui/file-input';
 import { AutoFormInputComponentProps } from '../type';
-import { AutoFormLabel } from './common/label';
-import { AutoFormTooltip } from './common/tooltip';
-import { AutoFormWrapper } from './common/wrapper';
 import { DefaultParent } from './common/children';
+import { AutoFormLabel } from './common/label';
+import { AutoFormWrapper } from './common/wrapper';
+import { FormControl, FormMessage } from '@/components/ui/form';
+import { AutoFormTooltip } from './common/tooltip';
 
-import { FormControl, FormMessage } from '../../form';
-import { ColorPicker } from '../../color-picker';
-
-export const AutoFormColor = ({
+export const AutoFormFile = ({
   autoFormProps: { isRequired, fieldConfigItem, field, theme, isDisabled },
   ...props
 }: AutoFormInputComponentProps &
-  Omit<
-    React.ComponentProps<typeof ColorPicker>,
-    'onChange' | 'required' | 'value'
-  >) => {
-  const value = field.value || '';
+  Omit<React.ComponentProps<typeof FileInput>, 'onChange' | 'value'>) => {
   const ParentWrapper = fieldConfigItem.renderParent ?? DefaultParent;
 
   return (
@@ -30,12 +25,10 @@ export const AutoFormColor = ({
       )}
       <ParentWrapper field={field}>
         <FormControl>
-          <ColorPicker
-            required={isRequired}
+          <FileInput
             {...field}
             {...props}
             disabled={isDisabled || props.disabled}
-            value={value}
           />
         </FormControl>
       </ParentWrapper>
