+++ packages/frontend/src/components/form/fields/icon.tsx
@@ -1,17 +1,17 @@
+import { IconPicker } from '@/components/icon/picker/icon-picker';
 import { AutoFormInputComponentProps } from '../type';
-import { AutoFormLabel } from './common/label';
-import { AutoFormTooltip } from './common/tooltip';
-import { AutoFormWrapper } from './common/wrapper';
 import { DefaultParent } from './common/children';
+import { AutoFormWrapper } from './common/wrapper';
+import { AutoFormLabel } from './common/label';
+import { FormControl, FormMessage } from '@/components/ui/form';
+import { AutoFormTooltip } from './common/tooltip';
 
-import { FormControl, FormMessage } from '../../form';
-import { FileInput } from '../../file-input';
-
-export const AutoFormFile = ({
+export const AutoFormIcon = ({
   autoFormProps: { isRequired, fieldConfigItem, field, theme, isDisabled },
   ...props
 }: AutoFormInputComponentProps &
-  Omit<React.ComponentProps<typeof FileInput>, 'onChange' | 'value'>) => {
+  Omit<React.ComponentProps<typeof IconPicker>, 'onChange' | 'value'>) => {
+  const value = field.value || '';
   const ParentWrapper = fieldConfigItem.renderParent ?? DefaultParent;
 
   return (
@@ -26,9 +26,10 @@ export const AutoFormFile = ({
       )}
       <ParentWrapper field={field}>
         <FormControl>
-          <FileInput
-            {...field}
+          <IconPicker
             {...props}
+            {...field}
+            value={value}
             disabled={isDisabled || props.disabled}
           />
         </FormControl>
