+++ packages/frontend/src/components/form/fields/text-language-input.tsx
@@ -1,36 +1,36 @@
+import { TextLanguageInput } from '@/components/ui/text-language-input';
 import { AutoFormInputComponentProps } from '../type';
-import { AutoFormLabel } from './common/label';
-import { AutoFormTooltip } from './common/tooltip';
-import { AutoFormWrapper } from './common/wrapper';
 import { DefaultParent } from './common/children';
+import { AutoFormWrapper } from './common/wrapper';
+import { AutoFormLabel } from './common/label';
+import { FormControl, FormMessage } from '@/components/ui/form';
+import { AutoFormTooltip } from './common/tooltip';
 
-import { FormControl, FormMessage } from '../../form';
-import { TagsInput } from '../../tags-input';
-
-export const AutoFormTags = ({
+export const AutoFormTextLanguageInput = ({
   autoFormProps: { isRequired, fieldConfigItem, field, theme, isDisabled },
   ...props
 }: AutoFormInputComponentProps &
-  Omit<React.ComponentProps<typeof TagsInput>, 'onChange'>) => {
-  const value = field.value || '';
+  Omit<
+    React.ComponentProps<typeof TextLanguageInput>,
+    'onChange' | 'value'
+  >) => {
   const ParentWrapper = fieldConfigItem.renderParent ?? DefaultParent;
 
   return (
     <AutoFormWrapper theme={theme}>
       {fieldConfigItem?.label && (
         <AutoFormLabel
-          description={fieldConfigItem.description}
           label={fieldConfigItem.label}
           isRequired={isRequired}
           theme={theme}
+          description={fieldConfigItem.description}
         />
       )}
       <ParentWrapper field={field}>
         <FormControl>
-          <TagsInput
-            {...props}
+          <TextLanguageInput
             {...field}
-            value={props.value ?? value}
+            {...props}
             disabled={isDisabled || props.disabled}
           />
         </FormControl>
