+++ packages/frontend/src/views/admin/views/core/langs/table/actions/delete/content.tsx
@@ -10,8 +10,8 @@ import {
 } from '@/components/ui/alert-dialog';
 import { Button } from '@/components/ui/button';
 import { ShowCoreLanguages } from '@/graphql/types';
-import { AutoForm } from '@/components/ui/auto-form';
-import { AutoFormInput } from '@/components/ui/auto-form/fields/input';
+import { AutoForm } from '@/components/form/auto-form';
+import { AutoFormInput } from '@/components/form/fields/input';
 
 export const ContentDeleteActionsTableLangsCoreAdmin = ({
   code,
