+++ packages/frontend/src/views/admin/views/core/plugins/actions/create/form.tsx
@@ -1,7 +1,7 @@
 import { useTranslations } from 'next-intl';
 
-import { AutoFormInput } from '@/components/ui/auto-form/fields/input';
-import { AutoForm } from '@/components/ui/auto-form';
+import { AutoFormInput } from '@/components/form/fields/input';
+import { AutoForm } from '@/components/form/auto-form';
 import { useCreateEditPluginAdmin } from './hooks/use-create-edit-plugin-admin';
 import { ShowAdminPlugins } from '@/graphql/types';
