+++ packages/frontend/src/views/admin/views/core/settings/main/content.tsx
@@ -4,9 +4,9 @@ import { useTranslations } from 'next-intl';
 
 import { useSettingsCoreAdmin } from './hooks/use-settings-core-admin';
 import { Core_Main_Settings__ShowQuery } from '@/graphql/queries/admin/settings/core_main_settings__show.generated';
-import { AutoForm } from '@/components/ui/auto-form';
-import { AutoFormInput } from '@/components/ui/auto-form/fields/input';
-import { AutoFormTextLanguageInput } from '@/components/ui/auto-form/fields/text-language-input';
+import { AutoForm } from '@/components/form/auto-form';
+import { AutoFormInput } from '@/components/form/fields/input';
+import { AutoFormTextLanguageInput } from '@/components/form/fields/text-language-input';
 
 export const ContentMainSettingsCoreAdmin = (
   props: Core_Main_Settings__ShowQuery,
